requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10
gunicorn==19.2.0
itsdangerous==0.24
python-firebase==1.2
requests==2.5.1
